WIP for test data
This adds a json environment variable that maps postcodes to tiers. It
should be passed in from concourse.

When the postcode is updated it is checked to see if it is in the list
of test postcodes

TODO:

Make work with address-lookup
Write tests
Filter out any submissions with the test data set
Make behave tests use the appropriate test postcodes
Make gatling tests use the appropriate test postcodes
Update concourse